Extracting runAndRetry logic
